Add ragel scanner mode support
Essentially the same but shifting is done on ts rather than p. Also note
that shifting is done when ts is non-zero, rather than half-way through
the data array. This is because a token that starts in the first half of
the data array will block advancement. This failure would probably
indicate a poorly chosen block size in the real world and also reflects
the fact that the fsm simulation in the tests does not accept completed
tokens. Probably worth fixing this to allow the more efficient copying
approach.